PD-220: Refactor RadioBox and RadioBoxGroup to eliminate console errors (#58)
* refactor RadioBox and RadioBoxGroup to eliminate console errors

* Stylistic tweaks to RadioBox

* export default inline with function declaration
